// app/data/enums.ts
export enum Ferramentas {
    KitDeArrombar = "Kit de Arrombar",
    KitDeHerborismo = "Kit de Herborismo",
    KitDeDisfarce = "Kit de Disfarce",
    KitDeVeneno = "Kit de Envenenamento",
    FerramentasDeLadrao = "Ferramentas de Ladrão",
    FerramentasDeArtesaoAlfaiate = "Ferramentas de Alfaiate",
    FerramentasDeArtesaoCeramica = "Ferramentas de Cerâmica",
    FerramentasDeArtesaoFerreiro = "Ferramentas de Ferreiro",
    FerramentasDeArtesaoMarceneiro = "Ferramentas de Marceneiro",
    FerramentasDeArtesaoPintor = "Ferramentas de Pintor",
    FerramentasDeArtesaoVidreiro = "Ferramentas de Vidreiro",
    FerramentasDeNavegacao = "Ferramentas de Navegação",
    FerramentasDeEscalada = "Ferramentas de Escalada",
    FerramentasDeFermentacao = "Ferramentas de Fermentação",
    FerramentasDeJoalheria = "Ferramentas de Joalheria",
    FerramentasDeAlquimia = "Ferramentas de Alquimia",
    JogoDeCartas = "Jogo de Cartas",
    JogoDeXadrez = "Jogo de Xadrez",
    JogoDeDados = "Jogo de Dados",
    FerramentasDeInstrumentoLira = "Lira",
    FerramentasDeInstrumentoFlauta = "Flauta",
    FerramentasDeInstrumentoTambor = "Tambor",
    FerramentasDeInstrumentoCítara = "Cítara",
    FerramentasDeInstrumentoGaita = "Gaita"
}
export enum Idiomas {
    Abissal = "Abissal",
    Anao = "Anão",
    Celestial = "Celestial",
    Comum = "Comum",
    Draconico = "Dracônico",
    Enanico = "Anão",
    Elfico = "Élfico",
    Gigante = "Gigante",
    Gnoll = "Gnoll",
    Goblin = "Goblin",
    Halfling = "Halfling",
    Infernal = "Infernal",
    Orc = "Orc",
    Primordial = "Primordial",
    Silvano = "Silvano",
    Subcomum = "Subcomum",
    Umbral = "Umbral"
}
export enum Pericias {
    Acrobacia,
    AdestrarAnimais,
    Arcanismo,
    Atletismo,
    Atuacao,
    Enganacao,
    Furtividade,
    Historia,
    Intimidacao,
    Intuicao,
    Investigacao,
    Medicina,
    Natureza,
    Percepcao,
    Persuasao,
    Prestidigitacao,
    Religiao,
    Sobrevivencia
}